chore: Move ConversationRenamedSystemMessageCell into its own file
